+++ .env.example
@@ -1,1 +1,6 @@
 SERVER_PORT=
+
+USER_DB=
+PASS_DB=
+
+PRIVATE_KEY=
