--- pathEnv.js
@@ -1,3 +1,0 @@
-module.exports = {
-	path: "/home/allyson/Área de Trabalho/Projetos Back-End/auth_users/.env",
-};
