+++ src/index.js
@@ -1,6 +1,5 @@
 const app = require("./server/server");
-const pathEnv = require("../pathEnv");
-require("dotenv").config(pathEnv);
+require("dotenv").config();
 
 const port = process.env.SERVER_PORT || 3000;
